move SetPasswordPage.js to new folder
