+++ src/pages/setpassword/SetPasswordPage.js
@@ -9,12 +9,12 @@ import PropTypes from 'prop-types';
 import {withOnyx} from 'react-native-onyx';
 import _ from 'underscore';
 import lodashGet from 'lodash.get';
-import styles from '../styles/styles';
-import ExpensifyCashLogo from '../../assets/images/expensify-cash.svg';
-import {setPassword} from '../libs/actions/Session';
-import ONYXKEYS from '../ONYXKEYS';
-import variables from '../styles/variables';
-import ButtonWithLoader from '../components/ButtonWithLoader';
+import styles from '../../styles/styles';
+import ExpensifyCashLogo from '../../../assets/images/expensify-cash.svg';
+import {setPassword} from '../../libs/actions/Session';
+import ONYXKEYS from '../../ONYXKEYS';
+import variables from '../../styles/variables';
+import ButtonWithLoader from '../../components/ButtonWithLoader';
 
 const propTypes = {
     /* Onyx Props */
